Fix parameter copy ctor
Closes #19 
Many thanks for the fix and the follow-up tests

+++ src/lcfiplus.cc
@@ -1279,7 +1279,7 @@ vector<const Vertex*> Jet::getVerticesForFT() const {
 // under construction
 void Parameters::remove(const string& key, bool delmap) {
   if (_map.find(key) == _map.end())throw(Exception("Parameters::remove: key not found."));
-  else if (_map[key].first == &typeid(double))deleteData<double>(key);
+  else if (_map[key].first == &typeid(float))deleteData<float>(key);
   else if (_map[key].first == &typeid(double))deleteData<double>(key);
   else if (_map[key].first == &typeid(long))deleteData<long>(key);
   else if (_map[key].first == &typeid(int))deleteData<int>(key);
@@ -1318,7 +1318,7 @@ Parameters& Parameters::operator =(const Parameters& ref) {
 
     // built-in types
     if (type == &typeid(double))add(key,*(double*)data);
-    else if (type == &typeid(double))add(key,*(double*)data);
+    else if (type == &typeid(float))add(key,*(float*)data);
     else if (type == &typeid(long))add(key,*(long*)data);
     else if (type == &typeid(int))add(key,*(int*)data);
     else if (type == &typeid(bool))add(key,*(bool*)data);
